insider_blog_microservices/requirements.txt
autopep8==1.6.0
cffi==1.15.1
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
distlib==0.3.4
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.5.0
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
virtualenv==20.14.1
Werkzeug==2.1.2
